Don't fail if there are no favorite posts
If there are no posts returned by `get_favorite_posts` the test would fail. 

When there are zero favorite posts, `get_favorite_posts` returns and empty array, which would pass the test for existence but would fail to have a `x[0]` item.

This tests instead for an empty array and skips assigning a favorite post if there isn't one.

This has not been tested, but should do the trick.